# docs/gtm-api-openapi.yaml
openapi: 3.0.3
info:
  title: Global Travel Monitor Events API
  description: |
    Global Travel Monitor (GTM) API – Read-only Zugriff auf globale
    Sicherheits- und Reiserisiko-Events sowie Länder-Übersichten.
  version: 1.0.0
  contact:
    name: Global Travel Monitor Support
servers:
  - url: https://platform.passolution.de/api/v1
    description: Produktionsserver (empfohlen)
  - url: https://api.global-travel-monitor.de/v1
    description: Bisherige Adresse, weiterhin gültig
  - url: https://global-travel-monitor.eu/api/v1
    description: Bisherige Adresse, weiterhin gültig
security:
  - BearerAuth: []
...
paths:
  /events:
    get:
      tags: [Events]
      summary: Events auflisten
      description: |
        Gibt eine paginierte Liste aller aktiven Events zurück (einschließlich
        zukünftiger Events). Dies umfasst sowohl von Global Travel Monitor gepflegte
        Events als auch Events von API-Partnern. Es werden nur Events zurückgegeben,
        die freigegeben (approved), aktiv und nicht archiviert sind. Mit start_date
        und end_date kann der Zeitraum eingegrenzt werden.
      parameters:
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [high, medium, low, info]
          description: Filter nach Risikostufe
        - name: country
          in: query
          schema:
            type: string
            maxLength: 3
          description: "Filter nach Ländercode – ISO alpha-2 (z.B. DE) oder alpha-3 (z.B. DEU)"
        - name: event_category
          in: query
          schema:
            type: string
            maxLength: 50
          description: "Filter nach Event-Typ-Code (z.B. safety)"
        - name: region
          in: query
          schema:
            type: integer
          description: Filter nach Region-ID (numerische ID)
        - name: source
          in: query
          schema:
            type: string
            maxLength: 100
          description: "Filter nach Event-Herkunft – Name des Anbieters (z.B. manual, API-Client-Name)"
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: "Nur Events ab diesem Datum (z.B. 2026-03-01)"
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: "Nur Events bis zu diesem Datum (z.B. 2026-04-30)"
        - name: per_page
          in: query
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
          description: Einträge pro Seite
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Seitennummer
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /events/countries:
    get:
      tags: [Events]
      summary: Länder mit aktiven Events auflisten
      description: |
        Gibt eine Liste aller Länder zurück, die mindestens ein aktives Event haben,
        zusammen mit der Anzahl aktiver Events. Sortiert nach Anzahl (absteigend).
        Nicht paginiert.
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CountryWithEvents'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /continents:
    get:
      tags: [Basisdaten]
      summary: Kontinente auflisten
      description: Gibt eine Liste aller Kontinente zurück.
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContinentItem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /countries:
    get:
      tags: [Basisdaten]
      summary: Alle Länder auflisten
      description: |
        Gibt eine Liste aller Länder zurück. Optional nach Kontinent filterbar.
      parameters:
        - name: continent
          in: query
          schema:
            type: string
            maxLength: 2
          description: "Filter nach Kontinent-Code (z.B. EU, AS, AF)"
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Country'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /countries/{code}:
    get:
      tags: [Basisdaten]
      summary: Alle Angaben eines Landes
      description: |
        Liefert Grunddaten, Länderbeschreibung, Reiseinformationen, Strom, Trinkgeld,
        Taxi-Apps, Mobilfunkanbieter, Feiertage, Bilder und Risikoprofil eines Landes.
        Mehrsprachige Texte kommen als Objekt je Sprache; mit `lang` als String in dieser Sprache.
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z]{2,3}$'
          description: "ISO 3166-1 alpha-2 oder alpha-3 Code (z.B. DE oder DEU)"
        - name: lang
          in: query
          schema:
            type: string
            enum: [de, en, nl]
          description: "Nur diese Sprache statt aller Sprachen"
        - name: year
          in: query
          schema:
            type: integer
            minimum: 1970
            maximum: 2100
          description: "Jahr der Feiertage (Standard: laufendes Jahr)"
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/CountryDetail'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /countries/{code}/boundary:
    get:
      tags: [Basisdaten]
      summary: Landesgrenze eines Landes als GeoJSON-Feature
      description: Auf rund 5 km vereinfachte Grenze (Quelle Natural Earth), Koordinaten als [lng, lat].
      parameters:
        - name: code
          in: path
          required: true
          schema: {type: string, example: EG}
        - name: lang
          in: query
          required: false
          schema: {type: string, enum: [de, en, nl]}
      responses:
        '200':
          description: Feature
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data: {$ref: '#/components/schemas/BoundaryFeature'}
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /countries/{code}/regions:
    get:
      summary: Regionen eines Landes
      description: Alle Regionen eines Landes, alphabetisch, mit Anzahl der zugeordneten Städte.
      tags: [Basisdaten]
      parameters:
        - name: code
          in: path
          required: true
          schema: {type: string}
          description: ISO-3166-1-Code (2 oder 3 Buchstaben)
        - name: lang
          in: query
          schema: {type: string, enum: [de, en, nl]}
          description: Nur diese Sprache statt aller Sprachen
      responses:
        "200":
          description: Regionen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/CountryRegion'}
                  meta: {type: object, properties: {total: {type: integer}}}
        "404": {description: Land nicht gefunden}
...
  /sights/{id}:
    get:
      summary: Einzelne Sehenswürdigkeit
      tags: [Basisdaten]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
        - name: lang
          in: query
          schema: {type: string, enum: [de, en, nl]}
          description: Nur diese Sprache statt aller Sprachen
      responses:
        "200":
          description: Sehenswürdigkeit mit Land
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data:
                    allOf:
                      - {$ref: '#/components/schemas/Sight'}
                      - type: object
                        properties:
                          country:
                            type: object
                            properties:
                              code: {type: string}
                              name: {description: "Text je Sprache oder – mit ?lang – ein String"}
        "404": {description: Sehenswürdigkeit nicht gefunden}
  /countries/{code}/weather:
    get:
      summary: Wetter der Hauptstadt eines Landes
      description: Aktuelles Wetter, 24 Stunden in 3-Stunden-Schritten und bis zu 5 Tage (OpenWeatherMap). Hauptstadt, sonst Landesmitte.
      tags: [Basisdaten]
      parameters:
        - name: code
          in: path
          required: true
          schema: {type: string}
        - name: lang
          in: query
          schema: {type: string, enum: [de, en, nl]}
      responses:
        "200":
          description: Wetterbericht
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data: {$ref: '#/components/schemas/WeatherReport'}
        "404": {description: Land nicht gefunden oder ohne Koordinaten}
        "503": {description: Wetterquelle nicht erreichbar}
...
  /airports/{code}:
    get:
      tags: [Basisdaten]
      summary: Flughafen mit Lounges, Hotels, Mobilität und Airlines
      parameters:
        - {name: code, in: path, required: true, description: IATA (3) oder ICAO (4), schema: {type: string, example: CAI}}
        - {name: lang, in: query, required: false, schema: {type: string, enum: [de, en, nl]}}
      responses:
        '200':
          description: Flughafen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data: {$ref: '#/components/schemas/AirportDetail'}
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /airlines/{code}:
    get:
      tags: [Basisdaten]
      summary: Airline mit Kontakt, Gepäck, Tierregelung und Flughäfen
      parameters:
        - {name: code, in: path, required: true, description: IATA (2) oder ICAO (3), schema: {type: string, example: MS}}
        - {name: include, in: query, required: false, description: "lounges, hotels (kommagetrennt) je Flughafen", schema: {type: string, example: "lounges,hotels"}}
        - {name: lang, in: query, required: false, schema: {type: string, enum: [de, en, nl]}}
      responses:
        '200':
          description: Airline
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data: {$ref: '#/components/schemas/AirlineDetail'}
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /exchange-rates:
    get:
      tags: [Basisdaten]
      summary: Aktuelle Wechselkurse
      description: Stündlich aktualisierte Kurse (Basis EUR, andere Basis als Kreuzkurs). `source` ist bei Anzeige zu nennen.
      parameters:
        - name: base
          in: query
          schema: {type: string, pattern: '^[A-Za-z]{3}$', default: EUR}
        - name: symbols
          in: query
          schema: {type: string}
          description: "Nur diese Währungen, kommagetrennt"
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  data:
                    type: object
                    properties:
                      base: {type: string, example: EUR}
                      rates:
                        type: object
                        additionalProperties: {type: number}
                        example: {EGP: 54.21, USD: 1.08}
                      updated_at: {type: string, format: date-time, nullable: true}
                      next_update_at: {type: string, format: date-time, nullable: true}
                      source: {type: string}
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          description: Anbieter nicht erreichbar und keine Kurse im Cache
  /regions:
    get:
      tags: [Basisdaten]
      summary: Regionen auflisten
      description: |
        Gibt eine Liste aller Regionen zurück.
        Optional nach Land filterbar.
      parameters:
        - name: country
          in: query
          schema:
            type: string
            maxLength: 3
          description: "Filter nach Ländercode – ISO alpha-2 (z.B. DE) oder alpha-3 (z.B. DEU)"
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Region'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /event-categories:
    get:
      tags: [Basisdaten]
      summary: Verfügbare Event-Typen auflisten
      description: |
        Gibt eine Liste aller aktiven Event-Typen zurück.
        Nützlich um die gültigen Werte für den event_category-Filter zu ermitteln.
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventCategory'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API-Token im Format \"Bearer {token}\""
  schemas:
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: UUID des Events
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Earthquake in Turkey
        description:
          type: string
          nullable: true
          example: A 6.2 magnitude earthquake struck southeastern Turkey.
        risk_level:
          type: string
          enum: [high, medium, low, info]
          example: high
        start_date:
          type: string
          format: date-time
          nullable: true
          example: "2025-03-15T08:30:00Z"
        end_date:
          type: string
          format: date-time
          nullable: true
          description: "null = andauernd"
          example: null
        latitude:
          type: number
          format: float
          nullable: true
          example: 37.7749
        longitude:
          type: number
          format: float
          nullable: true
          example: 35.3214
        is_nationwide:
          type: boolean
          description: "Landesweite Geltung: /events/nearby liefert das Event unabhängig vom Radius, sobald der Abfragepunkt in einem der betroffenen Länder liegt"
          example: false
        event_categories:
          type: array
          items:
            $ref: '#/components/schemas/EventCategory'
        countries:
          type: array
          items:
            $ref: '#/components/schemas/EventCountry'
        locations:
          type: array
          description: Alle Standorte des Events mit Land, Region, Stadt und Koordinaten
          items:
            $ref: '#/components/schemas/EventLocation'
        source:
          $ref: '#/components/schemas/EventSource'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CountryRegion:
      type: object
      properties:
        id: {type: integer, example: 10}
        code: {type: string, nullable: true, example: DE-BW}
        name: {description: "Text je Sprache oder – mit ?lang – ein String", example: {de: Baden-Württemberg, en: Baden-Württemberg}}
        description: {type: string, nullable: true}
        is_popular: {type: boolean, description: "Markierung „beliebt“ aus der Verwaltung"}
        coordinates:
          type: object
          nullable: true
          properties: {lat: {type: number}, lng: {type: number}}
        cities_count: {type: integer, example: 8}
...
    Sight:
      type: object
      properties:
        id: {type: integer, example: 3}
        name: {description: "Text je Sprache oder – mit ?lang – ein String", example: {de: Aquarium Genua, en: Aquarium of Genoa}}
        category: {type: string, example: theme_park}
        category_name: {description: "Bezeichnung der Kategorie je Sprache oder – mit ?lang – ein String"}
        is_highlight: {type: boolean}
        region:
          type: object
          nullable: true
          properties:
            id: {type: integer}
            code: {type: string, nullable: true}
            name: {description: "Text je Sprache oder – mit ?lang – ein String"}
        city:
          type: object
          nullable: true
          properties:
            id: {type: integer}
            name: {description: "Text je Sprache oder – mit ?lang – ein String"}
        coordinates:
          type: object
          nullable: true
          properties: {lat: {type: number}, lng: {type: number}}
        address: {type: string, nullable: true}
        website_url: {type: string, nullable: true}
        ticket_url: {type: string, nullable: true}
        visit_minutes: {type: integer, nullable: true, description: Übliche Besuchsdauer}
        short_description: {nullable: true, description: "Text je Sprache oder – mit ?lang – ein String"}
        description: {nullable: true, description: "Text je Sprache oder – mit ?lang – ein String"}
        opening_hours: {nullable: true, description: "Allgemein und ohne Gewähr"}
        admission: {nullable: true, description: "Allgemein und ohne Gewähr"}
        best_time: {nullable: true}
        tips: {nullable: true}
        accessibility: {nullable: true}
        is_reviewed: {type: boolean, description: "false, solange ein KI-Entwurf nicht redaktionell geprüft ist"}
    WeatherReport:
      type: object
      properties:
        location:
          type: object
          properties:
            name: {type: string, nullable: true, example: Berlin}
            lat: {type: number}
            lng: {type: number}
            timezone_offset: {type: integer, description: Versatz zu UTC in Sekunden, example: 7200}
        current:
          type: object
          properties:
            observed_at: {type: string, format: date-time}
            temperature: {type: number, description: "°C"}
            feels_like: {type: number}
            temp_min: {type: number, nullable: true}
            temp_max: {type: number, nullable: true}
            humidity: {type: integer, description: "%"}
            pressure: {type: integer, description: hPa}
            wind_speed: {type: number, description: km/h}
            wind_direction: {type: integer, description: Grad}
            visibility: {type: number, description: km}
            clouds: {type: integer, description: "%"}
            description: {type: string, example: Bedeckt}
            condition: {type: string, example: Bewölkt}
            icon: {type: string, example: 04n}
            icon_url: {type: string}
            sunrise: {type: string, format: date-time, nullable: true}
            sunset: {type: string, format: date-time, nullable: true}
        hourly:
          type: array
          description: Nächste 24 Stunden in 3-Stunden-Schritten
          items:
            type: object
            properties:
              time: {type: string, format: date-time}
              temperature: {type: number}
              icon: {type: string}
              description: {type: string}
              precipitation_probability: {type: integer, description: "%"}
              wind_speed: {type: number, description: km/h}
        daily:
          type: array
          description: Bis zu 5 Tage
          items:
            type: object
            properties:
              date: {type: string, format: date}
              temp_min: {type: number}
              temp_max: {type: number}
              icon: {type: string}
              description: {type: string}
              precipitation_probability: {type: integer, description: "%"}
        source:
          type: object
          properties:
            name: {type: string, example: OpenWeatherMap}
            url: {type: string}
...
    EventCategory:
      type: object
      properties:
        code:
          type: string
          example: safety
        name:
          type: string
          example: Sicherheit
...
    ContinentItem:
      type: object
      properties:
        code:
          type: string
          example: EU
        name_de:
          type: string
          example: Europa
        name_en:
          type: string
          example: Europe
        lat:
          type: number
          format: float
          nullable: true
          example: 54.526
        lng:
          type: number
          format: float
          nullable: true
          example: 15.2551
...
    Country:
      type: object
      properties:
        iso_code:
          type: string
          example: DE
        iso3_code:
          type: string
          example: DEU
        name_de:
          type: string
          example: Deutschland
        name_en:
          type: string
          example: Germany
        continent:
          type: string
          nullable: true
          example: Europe
        continent_de:
          type: string
          nullable: true
          example: Europa
        lat:
          type: number
          format: float
          nullable: true
          example: 51.1657
        lng:
          type: number
          format: float
          nullable: true
          example: 10.4515
        is_eu_member:
          type: boolean
          example: true
        is_schengen_member:
          type: boolean
          example: true
...
    AirportDetail:
      allOf:
        - $ref: '#/components/schemas/AirportSummary'
        - type: object
          properties:
            altitude_m: {type: integer, nullable: true}
            lounges:
              type: array
              items: {$ref: '#/components/schemas/Lounge'}
            nearby_hotels:
              type: array
              items: {$ref: '#/components/schemas/NearbyHotel'}
            mobility:
              type: object
              properties:
                taxi: {type: object, properties: {available: {type: boolean}, info: {type: string, nullable: true}, approx_cost: {type: string, nullable: true}}}
                parking: {type: object, properties: {available: {type: boolean}, options: {type: array, items: {$ref: '#/components/schemas/NamedLink'}}}}
                car_rental: {type: object, properties: {available: {type: boolean}, providers: {type: array, items: {$ref: '#/components/schemas/NamedLink'}}}}
                airport_shuttle: {type: object, properties: {available: {type: boolean}, info: {type: string, nullable: true}, url: {type: string, nullable: true}}}
                public_transport: {type: object, properties: {available: {type: boolean}, types: {type: array, items: {$ref: '#/components/schemas/NamedLink'}}}}
            airlines:
              type: array
              items:
                type: object
                properties:
                  iata_code: {type: string, nullable: true}
                  icao_code: {type: string, nullable: true}
                  name: {type: string}
                  terminal: {type: string, nullable: true}
                  direction: {type: string, nullable: true}
                  cabin_classes:
                    type: array
                    items: {$ref: '#/components/schemas/CabinClass'}
            updated_at: {type: string, format: date-time, nullable: true}
...
    BoundaryFeature:
      type: object
      properties:
        type: {type: string, example: Feature}
        properties:
          type: object
          properties:
            iso_code: {type: string}
            iso3_code: {type: string, nullable: true}
            name: {$ref: '#/components/schemas/TranslatedText'}
        bbox:
          type: array
          nullable: true
          description: "[min_lng, min_lat, max_lng, max_lat]"
          items: {type: number}
        geometry:
          type: object
          description: GeoJSON Polygon oder MultiPolygon, Koordinaten als [lng, lat]
...
    CountryWithEvents:
      type: object
      properties:
        iso_code:
          type: string
          example: DE
        iso3_code:
          type: string
          example: DEU
        name_de:
          type: string
          example: Deutschland
        name_en:
          type: string
          example: Germany
        continent:
          type: string
          nullable: true
          example: Europe
        continent_de:
          type: string
          nullable: true
          example: Europa
        lat:
          type: number
          format: float
          nullable: true
          example: 51.1657
        lng:
          type: number
          format: float
          nullable: true
          example: 10.4515
        is_eu_member:
          type: boolean
          example: true
        is_schengen_member:
          type: boolean
          example: true
        active_events_count:
          type: integer
          example: 3
...
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 25
        total:
          type: integer
          example: 142
        last_page:
          type: integer
          example: 6
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
...
  responses:
    Unauthenticated:
      description: Nicht authentifiziert – Token fehlt oder ungültig
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Unauthenticated.
    Forbidden:
      description: Zugriff verweigert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Token does not have the required permissions.
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Event not found
    ValidationError:
      description: Validierungsfehler
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    TooManyRequests:
      description: Rate Limit überschritten (60 Requests pro Minute)
      headers:
        Retry-After:
          schema:
            type: integer
          description: Sekunden bis zum nächsten erlaubten Request